Library Reference

DiagrammaticEquations.DiagrammaticEquations — Module

The DiagrammaticEquations module exports data structures which represent diagrammatic equations and functions which manipulate them.

DiagrammaticEquations.DecapodeACSet.decapodeacset.Decapode — Method

function Decapode(e::DecaExpr)

Takes a DecaExpr and returns a Decapode ACSet.

DiagrammaticEquations.DecapodeACSet.decapodeacset.SummationDecapode — Method

function SummationDecapode(e::DecaExpr)

Takes a DecaExpr and returns a SummationDecapode ACSet.

DiagrammaticEquations.Quantity — Type

DECQuantities in DiagrammaticEquations must be subtypes of Number to integrate with SymbolicUtils. An intermediary type, Quantity, makes it clearer that terms in the theory are "symbolic quantities" which behave like numbers. In the context of SymbolicUtils, a Number is any type that you can do arithmetic operations on.

DiagrammaticEquations.Rule — Type

Rule{T}

Rules are AbstractOperators which contain type signatures that may be used for type inference or function resolution.

DiagrammaticEquations.UserOperator — Type

UserOperator{T}

UserOperators are AbstractOperators which store unary or binary operators to be checked against Rules.

Catlab.WiringDiagrams.WiringDiagramAlgebras.oapply — Method

function oapply(relation::RelationDiagram, podes::Vector{D}) where {D<:OpenSummationDecapode}

Compose a list of Decapodes as specified by the given relation diagram.

The Decapodes must be given in the same order as they were specified in the relation.

State variables (such as the (C,V) given in the head of the following @relation) do not affect the result of a composition.

Examples

julia> compose_diff_adv = @relation (C,V) begin
  diffusion(C, ϕ₁)
  advection(C, ϕ₂, V)
  superposition(ϕ₁, ϕ₂, ϕ, C)
end;

julia> oapply(compose_diff_adv, [(Diffusion, [:C, :ϕ]),
  (Advection, [:C, :ϕ, :V]), (Superposition, [:ϕ₁, :ϕ₂, :ϕ, :C])]);

DiagrammaticEquations.Open — Method

Open(d::SummationDecapode{T,U,V}, names::AbstractVector{Symbol}) where {T,U,V}

creates an OpenSummationDecapode based on named variables rather than variable indices. See AlgebraicPetri.jl's Open for the analogous verion for LabelledReactionNetworks.

DiagrammaticEquations.average_rewrite — Method

function average_rewrite(d::SummationDecapode)

Compute each quantitity in the given Decapode by the average of all computation paths leading to that node.

DiagrammaticEquations.bundle! — Method

function bundle!(d::SummationDecapode)

Find repeated Op1, Op2, or Σ (summation) applications and merge them into single computations. This optimization removes redundant paths from a Decapode.

Bundling is applied repeatedly until convergence, since merging one kind of operation may expose duplicates of another kind.

This algorithm performs common subexpression elimination.

See also: bundle.

DiagrammaticEquations.bundle — Method

function bundle(d::SummationDecapode)

Return a copy of the given Decapode in which repeated Op1, Op2, or Σ (summation) applications are merged into single computations. This optimization removes redundant paths.

This algorithm performs common subexpression elimination.

See also: bundle!.

DiagrammaticEquations.bundle_op1s! — Method

function bundle_op1s!(d::SummationDecapode)

Find Op1 operations which share the same source variable and operator, and merge their targets. This removes redundant computations from the Decapode.

DiagrammaticEquations.bundle_op2s! — Method

function bundle_op2s!(d::SummationDecapode)

Find Op2 operations which share the same source variables and operator, and merge their result variables. This removes redundant computations from the Decapode.

DiagrammaticEquations.bundle_sums! — Method

function bundle_sums!(d::SummationDecapode)

Find Σ (summation) nodes whose sets of summand variables are identical, and merge them into a single summation. Only the case in which all summands are shared is handled; partial overlap is not merged.

DiagrammaticEquations.check_operator — Method

function check_operator(user::AbstractOperator, rule::Rule; check_name::Bool = false, ignore_nonapplicable_types::Bool = false)

Return the number of differences in types between a given user operator and a rule operator. If the rule does not apply to this operator, or they differ by a non-inferable type, the type difference is Inf.

The ignore_nonapplicable_types flags apply to the user operator.

DiagrammaticEquations.collate — Method

function collate(equations, boundaries, uwd, symbols; restrictions::Vector{Symbol})

Create a collage of two Decapodes that simulates with boundary conditions.

Passing a list of symbols into restrictions specifies a list of boundary variables that will be restricted.

DiagrammaticEquations.contract_operators — Method

function contract_operators(d::SummationDecapode; white_list::Set{Symbol} = Set{Symbol}(), black_list::Set{Symbol} = Set{Symbol}())

Find chains of Op1s in the given Decapode, and replace them with a single Op1 with a vector of function names. After this process, all Vars that are not a part of any computation are removed. If a white list is provided, only chain those operators. If a black list is provided, do not chain those operators.

DiagrammaticEquations.default_composition_diagram — Method

function default_composition_diagram(podes::Vector{D}, names::Vector{Symbol}) where {D<:SummationDecapode}

Given a list of Decapodes and their names, return a composition diagram which assumes that variables sharing the same name ought to be composed.

No Literals are exposed. Use unique_lits! after composing.

Throw an error if any individual Decapode already contains a repeated name (except for Literals).

If only_states_terminals is true, only expose state and terminal variables. Defaults to false.

Note that composing immediately with oapply will fail if types do not match (e.g. (:infer, :Form0) or (:Form0, :Form1)).

DiagrammaticEquations.dot_rename! — Method

dot_rename!(d::AbstractNamedDecapode)

Rename tangent variables by their depending variable appended with a dot. e.g. If D == ∂ₜ(C), then rename D to Ċ.

If a tangent variable updates multiple vars, choose one arbitrarily. e.g. If D == ∂ₜ(C) and D == ∂ₜ(B), then rename D to either Ċ or B ̇.

DiagrammaticEquations.expand_operators — Method

function expand_operators(d::AbstractNamedDecapode)

If any unary operator is a composition, expand it out using intermediate variables.

DiagrammaticEquations.expand_operators — Method

function expand_operators(d::SummationDecapode)

Find operations that are compositions, and expand them with intermediate variables.

DiagrammaticEquations.fill_names! — Method

function fill_names!(d::AbstractNamedDecapode; lead_symbol::Symbol = Symbol("•"))

Provide a variable name to all the variables that don't have names.

DiagrammaticEquations.filterfor_ec_types — Method

filterfor_ec_types(types::AbstractVector{Symbol})

Return any form or vector-field type symbols.

DiagrammaticEquations.find_chains — Method

function find_chains(d::SummationDecapode; white_list::Set{Symbol} = Set{Symbol}(), black_list::Set{Symbol} = Set{Symbol}())

Find chains of Op1s in the given Decapode. A chain ends when the target of the last Op1 is part of an Op2 or sum, or is a target of multiple Op1s. If a white list is provided, only chain those operators. If a black list is provided, do not chain those operators.

DiagrammaticEquations.find_dep_and_order — Method

find_dep_and_order(d::AbstractNamedDecapode)

Find the order of each tangent variable in the Decapode, and the index of the variable that it is dependent on. Returns a tuple of (dep, order), both of which respecting the order in which incident(d, :∂ₜ, :op1) returns Vars.

DiagrammaticEquations.find_tgts_of_many_ops — Method

function find_tgts_of_many_ops(d::SummationDecapode)

Searches SummationDecapode, d, for all Vars which have two or more distinct operations leading into the same variable.

DiagrammaticEquations.get_valid_op1s — Method

function get_valid_op1s(deca_source::SummationDecapode, varID)

Searches SummationDecapode, deca_source, at the request varID and returns all op1s which are allowed to be averaged. Returns an array of indices of valid op1 sources.

Namely this is meant to exclude ∂ₜ from being included in an average.

DiagrammaticEquations.infer_resolve! — Method

infer_resolve!(d::SummationDecapode, operators::AbstractVector{Rule{Symbol}})

Runs type inference, overload resolution and type checking in that order.

DiagrammaticEquations.infer_state_names — Method

function infer_state_names(d)

Find names of variables which have a time derivative or are not the source of a computation. See also: infer_states.

DiagrammaticEquations.infer_states — Method

function infer_states(d::SummationDecapode)

Find variables which have a time derivative or are not the source of a computation. See also: infer_terminals.

DiagrammaticEquations.infer_terminal_names — Method

function infer_terminal_names(d)

Find names of variables which have no children. See also: infer_terminals.

DiagrammaticEquations.infer_terminals — Method

function infer_terminals(d::SummationDecapode)

Find variables which have no children. See also: infer_states.

DiagrammaticEquations.infer_types! — Method

function infer_types!(d::SummationDecapode, op1_rules::Vector{NamedTuple{(:src_type, :tgt_type, :replacement_type, :op), NTuple{4, Symbol}}})

Infer types of Vars given rules wherein one type is known and the other not.

DiagrammaticEquations.is_expanded — Method

is_expanded(d::AbstractNamedDecapode)

Check that no unary operator is a composition of unary operators.

DiagrammaticEquations.is_tgt_of_many_ops — Method

function is_tgt_of_many_ops(d::SummationDecapode, var)

Return true if there are two or more distinct operations leading into Var var (not counting ∂ₜ).

DiagrammaticEquations.remove_neighborless_vars! — Method

function remove_neighborless_vars!(d::SummationDecapode)

Remove all Vars from the given Decapode that are not part of any computation.

DiagrammaticEquations.replace_all_op1s! — Method

function replace_all_op1s!(d::SummationDecapode, LHS::Union{Symbol, SummationDecapode}, RHS::Union{Symbol, SummationDecapode})

Given a Decapode, d, replace all instances of the left-hand-side unary operator with those of the right-hand-side.

Return true if any replacements were made, otherwise false.

See also: replace_op1!, replace_all_op2s!

DiagrammaticEquations.replace_all_op2s! — Method

function replace_all_op2s!(d::SummationDecapode, LHS::Union{Symbol, SummationDecapode}, RHS::Union{Symbol, SummationDecapode}, proj1::Int, proj2::Int)

Given a Decapode, d, replace all instances of the left-hand-side binary operator with those of the right-hand-side.

proj1 and proj2 are the indices of the intended proj1 and proj2 in RHS.

Return true if any replacements were made, otherwise false.

See also: replace_op2!, replace_all_op1s!

DiagrammaticEquations.replace_all_op2s! — Method

function replace_all_op2s!(d::SummationDecapode, LHS::Union{Symbol, SummationDecapode}, RHS::Union{Symbol, SummationDecapode})

Given a Decapode, d, replace all instances of the left-hand-side binary operator with those of the right-hand-side.

Search for distinguished variables "p1" and "p2" to serve as the proj1 and proj2 from RHS.

Return true if any replacements were made, otherwise false.

See also: replace_op2!, replace_all_op1s!

DiagrammaticEquations.replace_op1! — Method

function replace_op1!(d::SummationDecapode, LHS::SummationDecapode, RHS::SummationDecapode)

Given a Decapode, d, replace at most one instance of the left-hand-side unary operator with those of the right-hand-side.

Return the index of the replaced unary operator, 0 if no match was found. See also: replace_op2!, replace_all_op1s!

DiagrammaticEquations.replace_op1! — Method

function replace_op1!(d::SummationDecapode, LHS::Symbol, RHS::SummationDecapode)

Given a Decapode, d, replace at most one instance of the left-hand-side unary operator with those of the right-hand-side.

Return the index of the replaced operator, 0 if no match was found.

See also: replace_all_op1s!

DiagrammaticEquations.replace_op1! — Method

function replace_op1!(d::SummationDecapode, LHS::Symbol, RHS::Symbol)

Given a Decapode, d, replace at most one instance of the left-hand-side unary operator with that of the right-hand-side.

Return the index of the replaced unary operator, 0 if no match was found. See also: replace_op2!, replace_all_op1s!

DiagrammaticEquations.replace_op2! — Method

function replace_op2!(d::SummationDecapode, LHS::SummationDecapode, RHS::SummationDecapode, proj1::Int, proj2::Int)

Given a Decapode, d, replace at most one instance of the left-hand-side binary operator with those of the right-hand-side.

proj1 and proj2 are the indices of the intended proj1 and proj2 in RHS.

Return the index of the replaced binary operator, 0 if no match was found. See also: replace_op1!, replace_all_op2s!

DiagrammaticEquations.replace_op2! — Method

function replace_op2!(d::SummationDecapode, LHS::Symbol, RHS::SummationDecapode, proj1::Int, proj2::Int)

Given a Decapode, d, replace at most one instance of the left-hand-side binary operator with those of the right-hand-side.

proj1 and proj2 are the indices of the intended proj1 and proj2 in RHS.

Return the index of the replaced operator, 0 if no match was found.

See also: replace_op1!, replace_all_op2s!

DiagrammaticEquations.replace_op2! — Method

function replace_op2!(d::SummationDecapode, LHS::Symbol, RHS::Symbol)

Given a Decapode, d, replace at most one instance of the left-hand-side binary operator with that of the right-hand-side.

Return the index of the replaced binary operator, 0 if no match was found. See also: replace_op1!, replace_all_op2s!

DiagrammaticEquations.resolve_overloads! — Method

function resolve_overloads!(d::SummationDecapode, op1_rules::Vector{NamedTuple{(:src_type, :tgt_type, :resolved_name, :op), NTuple{4, Symbol}}})

Resolve function overloads based on types of src and tgt.

DiagrammaticEquations.safe_modifytype! — Method

safe_modifytype!(d::SummationDecapode, var_idx::Int, new_type::Symbol)

This function calls safe_modifytype to safely modify a Decapode's variable type.

DiagrammaticEquations.safe_modifytype — Method

safe_modifytype(org_type::Symbol, new_type::Symbol)

This function accepts an original type and a new type and determines if the original type can be safely overwritten by the new type.

DiagrammaticEquations.transfer_children! — Method

function transfer_children!(d::SummationDecapode, x, y)

Transfer the children of x to y.

DiagrammaticEquations.transfer_parents! — Method

function transfer_parents!(d::SummationDecapode, x, y)

Transfer the parents of x to y. Also transfer TVar status from x to y.

DiagrammaticEquations.type_check — Method

type_check(d::SummationDecapode, type_rules::AbstractVector{Rule{Symbol}})

Takes a Decapode and a set of rules and checks to see if the operators that are in the Decapode contain a valid configuration of input/output types. If an operator in the Decapode does not contain a rule in the rule set it will be seen as valid.

In the case of a type error a DecaTypeExeception is thrown. Otherwise true is returned.

DiagrammaticEquations.type_check_Decapodes_composition — Method

function type_check_Decapodes_composition(relation::RelationDiagram, decs::Vector{OpenSummationDecapode})

Check that the types of all Vars connected by the same junction match.

This function only throws an error on the first type mismatch found.

DiagrammaticEquations.unique_by! — Method

function unique_by!(acset, column_names::Vector{Symbol})

Given column names from the same table, remove duplicate rows.

WARNING: This function does not check if other tables index into the one given. Removal of rows is performed with prejudice.

See also: unique_by.

Examples

julia> unique_by!(parallel_arrows(Graph, 123), :E, [:src,:tgt]) == parallel_arrows(Graph, 1)
true

DiagrammaticEquations.unique_by — Method

function unique_by(acset, column_names::Vector{Symbol})

Given column names from the same table, return a copy of the acset with duplicate rows removed. Removal of rows is performed with prejudice.

WARNING: This function does not check if other tables index into the one given. Removal of rows is performed with prejudice.

See also: unique_by!.

Examples

julia> unique_by(parallel_arrows(Graph, 123), :E, [:src,:tgt]) == parallel_arrows(Graph, 1)
true

DiagrammaticEquations.unique_lits! — Method

function unique_lits!(d::SummationDecapode)

Remove repeated Literals from a Decapode.

DiagrammaticEquations.@alias — Macro

Given a tuple of symbols ("aliases") and their canonical name (or "rep"), produces for each alias typechecking and nameof methods which call those for their rep. Example: @alias (d₀, d₁) => d

DiagrammaticEquations.@decapode — Macro

macro decapode(e)

Construct a SummationDecapode using the Decapode Domain-Specific Language.

DiagrammaticEquations.@operator — Macro

Creates an operator foo with arguments which are types in a given Theory. This entails creating (1) a function which performs type construction and (2) a function which consumes BasicSymbolic variables and returns Terms.

@operator foo(S1, S2, ...)::Theory begin
    (body of function)
    (@rule expr1)
    ...
    (@rule exprN)
end

builds

promote_symtype(::typeof{f}, ::Type{S1}, ::Type{S2}, ...) where {S1<:DECQuantity, S2<:DECQuantity, ...}
    (body of function)
end

as well as

foo(S1, S2, ...) where {S1<:DECQuantity, ...}
    s = promote_symtype(f, S1, S2, ...)
    SymbolicUtils.Term{s}(foo, [S1, S2, ...])
end

Example:

@operator Δ(s)::DECQuantity begin
    @match s begin
        ::Scalar => error("Invalid")
        ::VField => error("Invalid")
        ::Form => ⋆(d(⋆(d(s))))
    end
    @rule ~s --> ⋆(d(⋆(d(~s))))
end

DiagrammaticEquations.Deca.op1_operators — Constant

These are the default rules used to do type inference/function resolution in the 1D/2D exterior calculus.

Catlab.Graphics.GraphvizGraphs.to_graphviz — Method

Graphics.to_graphviz(F::AbstractDecapode; directed = true, kw...)

Visualize the given Decapode through Graphviz. Ensure that you have called using Catlab.Graphics before-hand, and have a way of visualizing SVG files in your current environment.

DiagrammaticEquations.Deca.get_colors — Method

get_colors(d::SummationDecapode)

Given a Decapode, we infer the name of the boxes in the cospan which defines it and associate to these boxes a random unique color from a list of colors.

DiagrammaticEquations.Deca.recursive_delete_parents — Method

function recursive_delete_parents!(d::SummationDecapode, to_delete::Vector{Int64})

Delete the given nodes and their parents in the Decapode, recursively.

DiagrammaticEquations.Deca.unicode! — Method

function unicode!(d::SummationDecapode)

Replace ASCII operators with their Unicode equivalents.

DiagrammaticEquations.Deca.vec_to_dec! — Method

function vec_to_dec!(d::SummationDecapode)

Replace Vector Calculus operators with Discrete Exterior Calculus equivalents.